SetLength aangepast, meer token checks, unit tests

+++ src/main/java/nl/han/dea/jasmijn/PlayListController.java
@@ -1,11 +1,7 @@
 package nl.han.dea.jasmijn;
 
-
-import nl.han.dea.jasmijn.dao.PlayListDAO;
-import nl.han.dea.jasmijn.dao.TrackDAO;
 import nl.han.dea.jasmijn.dto.PlayListDTO;
 import nl.han.dea.jasmijn.dto.TrackDTO;
-import nl.han.dea.jasmijn.dto.TracksDTO;
 import nl.han.dea.jasmijn.services.PlayListService;
 import nl.han.dea.jasmijn.services.TrackService;
 import nl.han.dea.jasmijn.services.UserService;
@@ -14,72 +10,84 @@ import javax.inject.Inject;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.sql.SQLException;
-
 
 @Path("/playlists")
 public class PlayListController {
     private PlayListService playListService;
     private UserService userService;
     private TrackService trackService;
-    private PlayListDAO playListDAO;
 
     @GET
     @Produces({MediaType.APPLICATION_JSON})
     public Response showPlayLists(){
-        return Response.ok(playListService.all()).build();
+        return Response.ok(playListService.allPlayLists()).build();
     }
 
     @GET
     @Produces({MediaType.APPLICATION_JSON})
     @Path("{id}/tracks")
-    public Response showTracks(@PathParam("id") int playlistId, @QueryParam("token") String token){
+    public Response showTracks(@PathParam("id") int playListId, @QueryParam("token") String token){
         if(!userService.tokenIsCorrect(token)){
             return Response.status(401).build();
         }
-        TracksDTO tracksDTO = new TracksDTO(trackService.getTracksByPlaylistId(playlistId));
-        return Response.ok(tracksDTO).build();
-    }
-
-    @POST//Test commit
-    @Produces({MediaType.APPLICATION_JSON})
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Path("{id}/tracks")
-    public void addTrackToPlaylist(@PathParam("id") int playListId, TrackDTO trackDTO){
-        playListDAO.addTrackToPlayList(playListId,trackDTO);
-    }
-
-    @Path("{id}")
-    @PUT
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response updatePlayList(@PathParam("id") int id, PlayListDTO playListDTO) {
-        playListService.updatePlayList(id, playListDTO);
-        return Response.ok(playListService.all()).build();
+        return Response.ok(trackService.getTracksByPlaylistId(playListId)).build();
     }
 
     @POST
     @Produces({MediaType.APPLICATION_JSON})
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response createPlayList(PlayListDTO playListDTO){
+    @Path("{id}/tracks")
+    public Response addTrackToPlaylist(@PathParam("id") int playListId, TrackDTO trackDTO, @QueryParam("token") String token){
+        if(!userService.tokenIsCorrect(token)){
+            return Response.status(401).build();
+        }
+        playListService.addTrackToPlayList(playListId,trackDTO);
+        return Response.ok(trackService.getTracksByPlaylistId(playListId)).build();
+    }
+
+    @Path("{id}")
+    @PUT
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response updatePlayList(@PathParam("id") int id, PlayListDTO playListDTO, @QueryParam("token") String token) {
+        if(!userService.tokenIsCorrect(token)){
+            return Response.status(401).build();
+        }
+        playListService.updatePlayList(id, playListDTO);
+        return Response.ok(playListService.allPlayLists()).build();
+    }
+
+    @POST
+    @Produces({MediaType.APPLICATION_JSON})
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response createPlayList(PlayListDTO playListDTO, @QueryParam("token") String token){
+        if(!userService.tokenIsCorrect(token)){
+            return Response.status(401).build();
+        }
         playListService.createPlayList(playListDTO);
-        return Response.ok(playListService.all()).build();
+        return Response.ok(playListService.allPlayLists()).build();
     }
 
     @Path("{id}")
     @DELETE
     @Produces({MediaType.APPLICATION_JSON})
-    public Response deletePlayList(@PathParam("id") int id){
+    public Response deletePlayList(@PathParam("id") int id, @QueryParam("token") String token){
+        if(!userService.tokenIsCorrect(token)){
+            return Response.status(401).build();
+        }
         playListService.deletePlayList(id);
-        return Response.ok(playListService.all()).build();
+        return Response.ok(playListService.allPlayLists()).build();
     }
 
     @Path("{id}/tracks/{trackId}")
     @DELETE
     @Produces({MediaType.APPLICATION_JSON})
-    public void deleteTrackFromPlayList(@PathParam("id") int playListId, @PathParam("trackId") int trackId){
+    public Response deleteTrackFromPlayList(@PathParam("id") int playListId, @PathParam("trackId") int trackId, @QueryParam("token") String token){
+        if(!userService.tokenIsCorrect(token)){
+            return Response.status(401).build();
+        }
         playListService.deleteTrackFromPlayList(playListId, trackId);
+        return Response.ok(trackService.getTracksByPlaylistId(playListId)).build();
     }
-
 
     @Inject
     public void setUserService(UserService userService){
@@ -94,11 +102,6 @@ public class PlayListController {
     @Inject
     public void setTrackService(TrackService trackService){
         this.trackService = trackService;
-    }
-
-    @Inject
-    public void setPlayListDAO(PlayListDAO playListDAO){
-        this.playListDAO = playListDAO;
     }
 
 }
